requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
transformers>=4.21.0
pillow>=9.0.0
numpy>=1.24.0
torch>=2.0.0
